Chapter 2 (fix)
blah.

+++ CS_Chapter2/HoursAndMinutes/Program.cs
@@ -9,7 +9,7 @@ namespace HoursAndMinutes
     {
         static void Main(string[] args)
         {
-            int m = 197;
+            double m = 197;
             double h = m / 60;
             Console.WriteLine("{0} minutes is equal to {0} hours and {1} minutes.", m / 60, m % 60);
         }
